Move models to plugin folder
