+++ setup.cfg
@@ -63,12 +63,12 @@ where = src
 
 [options.entry_points]
 ewatercycle.models =
-    Hype = ewatercycle.models.hype:Hype
-    Lisflood = ewatercycle.models.lisflood:Lisflood
-    Wflow = ewatercycle.models.wflow:Wflow
-    PCRGlobWB = ewatercycle.models.pcrglobwb:PCRGlobWB
-    MarrmotM01 = ewatercycle.models.marrmot:MarrmotM01
-    MarrmotM14 = ewatercycle.models.marrmot:MarrmotM14
+    Hype = ewatercycle.plugins.hype.model:Hype
+    Lisflood = ewatercycle.plugins.lisflood.model:Lisflood
+    Wflow = ewatercycle.plugins.wflow.model:Wflow
+    PCRGlobWB = ewatercycle.plugins.pcrglobwb.model:PCRGlobWB
+    MarrmotM01 = ewatercycle.plugins.marrmot.model:MarrmotM01
+    MarrmotM14 = ewatercycle.plugins.marrmot.model:MarrmotM14
 
 [options.extras_require]
 dev =
