+++ src/ewatercycle/defaults/bmi_handles.py
@@ -1,0 +1,7 @@
+
+
+class BmiHandles:
+    """eWaterCycle BMI handles.
+
+    """
+
